test setup refactor
