+++ test/standalone/standalone.spec.ts
@@ -1,4 +1,4 @@
-import "../setup/checkConnection";
+import "../checkConnection";
 //
 import { describe } from "@jest/globals";
 import { standaloneAnyTests } from "./any";
