+++ test/statement.spec.ts
@@ -1,4 +1,4 @@
-import "./setup/checkConnection";
+import "./checkConnection";
 //
 import { describe, expect, it } from "@jest/globals";
 import { ComposableOverpassStatement, OverpassApiObject, OverpassState } from "../src";
